% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.evaluate.core.R
\name{box.evaluate.core}
\alias{box.evaluate.core}
\title{Box Plots}
\usage{
box.evaluate.core(data, names, quantitative, selected)
}
\arguments{
\item{data}{The data as a data frame object. The data frame should possess
one row per individual and columns with the individual names and multiple
trait/character data.}

\item{names}{Name of column with the individual names as a character string}

\item{quantitative}{Name of columns with the quantitative traits as a
character vector.}

\item{selected}{Character vector with the names of individuals selected in
core collection and present in the \code{names} column.}
}
\value{

}
\description{
Plot Box-and-Whisker plots
\insertCite{tukey_exploratory_1970,mcgill_variations_1978}{EvaluateCore} to
graphically compare the probability distributions of quantitative traits
between entire collection (EC) and core set (CS).
}
\examples{

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[graphics]{boxplot}}, \code{\link[ggplot2]{geom_boxplot}}
}
