% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.evaluate.core.R
\name{checks.evaluate.core}
\alias{checks.evaluate.core}
\title{Common checks for all functions}
\usage{
checks.evaluate.core(
  data,
  names,
  quantitative = NULL,
  qualitative = NULL,
  selected
)
}
\arguments{
\item{data}{The data as a data frame object. The data frame should possess
one row per individual and columns with the individual names and multiple
trait/character data.}

\item{names}{Name of column with the individual names as a character string}

\item{quantitative}{Name of columns with the quantitative traits as a
character vector.}

\item{qualitative}{Name of columns with the qualitative traits as a character
vector.}

\item{selected}{Character vector with the names of individuals selected in
core collection and present in the \code{names} column.}
}
\description{
Not exported. Strictly internal
}
\keyword{internal}
