% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr.evaluate.core.R
\name{cr.evaluate.core}
\alias{cr.evaluate.core}
\title{Coincidence Rate of Range}
\usage{
cr.evaluate.core(data, names, quantitative, selected)
}
\arguments{
\item{data}{The data as a data frame object. The data frame should possess
one row per individual and columns with the individual names and multiple
trait/character data.}

\item{names}{Name of column with the individual names as a character string}

\item{quantitative}{Name of columns with the quantitative traits as a
character vector.}

\item{selected}{Character vector with the names of individuals selected in
core collection and present in the \code{names} column.}
}
\value{

}
\description{
Compute the Coincidence Rate of Range (CR)
\insertCite{diwan_methods_1995,hu_methods_2000}{EvaluateCore} to compare
quantitative traits of the entire collection (EC) and core set (CS).
}
\examples{

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[stats]{wilcox.test}}
}
