% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfdist.evaluate.core.R
\name{pdfdist.evaluate.core}
\alias{pdfdist.evaluate.core}
\title{Distance Between Probability Distributions}
\usage{
pdfdist.evaluate.core(data, names, quantitative, selected)
}
\arguments{
\item{data}{The data as a data frame object. The data frame should possess
one row per individual and columns with the individual names and multiple
trait/character data.}

\item{names}{Name of column with the individual names as a character string}

\item{quantitative}{Name of columns with the quantitative traits as a
character vector.}

\item{selected}{Character vector with the names of individuals selected in
core collection and present in the \code{names} column.}
}
\value{

}
\description{
Compute Kullback–Leibler
\insertCite{kullback_information_1951}{EvaluateCore}, Kolmogorov-Smirnov
\insertCite{kolmogorov_sulla_1933,smirnov_table_1948}{EvaluateCore} and
Anderson-Darling Distances
\insertCite{anderson_asymptotic_1952}{EvaluateCore} between the probability
distributions of collection (EC) and core set (CS) for quantitative traits.
}
\examples{

}
\seealso{
\code{\link[entropy]{KL.plugin}}, \code{\link[stats]{ks.test}},
  \code{\link[kSamples]{ad.test}}
}
