% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon.evaluate.core.R
\name{shannon.evaluate.core}
\alias{shannon.evaluate.core}
\title{Shannon-Weaver Diversity Index}
\usage{
shannon.evaluate.core(data, names, qualitative, selected)
}
\arguments{
\item{data}{The data as a data frame object. The data frame should possess
one row per individual and columns with the individual names and multiple
trait/character data.}

\item{names}{Name of column with the individual names as a character string}

\item{qualitative}{Name of columns with the qualitative traits as a character
vector.}

\item{selected}{Character vector with the names of individuals selected in
core collection and present in the \code{names} column.}
}
\value{

}
\description{
Compute the Shannon-Weaver Diversity Index
(\ifelse{html}{\out{<i>H'</i>}}{\eqn{H'}}), Maximum diversity
(\ifelse{html}{\out{<i>H'<sub>max</sub></i>}}{\eqn{H'_{max}}}) and Shannon
Equitability Index (\ifelse{html}{\out{<i>E<sub>H</sub></i>}}{\eqn{E_{H}}})
\insertCite{shannon_mathematical_1949}{EvaluateCore} to compare the
phenotypic diversity for qualitative traits between entire collection (EC)
and core set (CS).
}
\examples{

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[psych:misc]{shannon}}
}
