% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar.evaluate.core.R
\name{bar.evaluate.core}
\alias{bar.evaluate.core}
\title{Bar Plots}
\usage{
bar.evaluate.core(data, names, qualitative, selected)
}
\arguments{
\item{data}{The data as a data frame object. The data frame should possess
one row per individual and columns with the individual names and multiple
trait/character data.}

\item{names}{Name of column with the individual names as a character string}

\item{qualitative}{Name of columns with the qualitative traits as a character
vector.}

\item{selected}{Character vector with the names of individuals selected in
core collection and present in the \code{names} column.}
}
\value{

}
\description{
Plot Bar plots to graphically compare the frequency distributions of
qualitative traits between entire collection (EC) and core set (CS).
}
\examples{

}
\seealso{
\code{\link[graphics]{barplot}}, \code{\link[ggplot2]{geom_bar}}
}
