% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisquare.evalueate.core.R
\name{chisquare.evaluate.core}
\alias{chisquare.evaluate.core}
\title{Chi-squared Test for Homogenity}
\usage{
chisquare.evaluate.core(data, names, qualitative, selected)
}
\arguments{
\item{data}{The data as a data frame object. The data frame should possess
one row per individual and columns with the individual names and multiple
trait/character data.}

\item{names}{Name of column with the individual names as a character string}

\item{qualitative}{Name of columns with the qualitative traits as a
character vector.}

\item{selected}{Character vector with the names of individuals selected in
core collection and present in the \code{names} column.}
}
\value{

}
\description{
Compare the distribution frequencies of qualitative traits between entire
collection (EC) and core set (CS) by Chi-squared test for homogenity.
}
\examples{

}
\seealso{
\code{\link[stats]{chisq.test}}
}
