% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.evaluate.core.R
\name{corr.evaluate.core}
\alias{corr.evaluate.core}
\title{Phenotypic Correlations}
\usage{
corr.evaluate.core(data, names, quantitative, qualitative, selected)
}
\arguments{
\item{data}{The data as a data frame object. The data frame should possess
one row per individual and columns with the individual names and multiple
trait/character data.}

\item{names}{Name of column with the individual names as a character string}

\item{quantitative}{Name of columns with the quantitative traits as a
character vector.}

\item{qualitative}{Name of columns with the qualitative traits as a character
vector.}

\item{selected}{Character vector with the names of individuals selected in
core collection and present in the \code{names} column.}
}
\value{

}
\description{
Compute phenotypic correlations \insertCite{pearson_note_1895}{EvaluateCore}
between traits, plot correlation matrices as correlograms
\insertCite{friendly_corrgrams_2002}{EvaluateCore} and calculate mantel
correlation \insertCite{legendre_interpretation_2012}{EvaluateCore} between
them to compare entire collection (EC) and core set (CS).
}
\examples{

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[stats]{cor}}, \code{\link[ggcorrplot:ggcorrplot]{cor_pmat}}
  \code{\link[ggcorrplot]{ggcorrplot}}, \code{\link[vegan]{mantel}}
}
