% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.evaluate.core.R
\name{dist.evaluate.core}
\alias{dist.evaluate.core}
\title{Distance Measures}
\usage{
dist.evaluate.core(data, names, quantitative, qualitative, selected)
}
\arguments{
\item{data}{The data as a data frame object. The data frame should possess
one row per individual and columns with the individual names and multiple
trait/character data.}

\item{names}{Name of column with the individual names as a character string}

\item{quantitative}{Name of columns with the quantitative traits as a
character vector.}

\item{qualitative}{Name of columns with the qualitative traits as a character
vector.}

\item{selected}{Character vector with the names of individuals selected in
core collection and present in the \code{names} column.}
}
\value{

}
\description{
Compute Average entry-to-nearest-entry distance, Average
accession-to-nearest-entry distance and Average entry-to-entry distance
\insertCite{odong_quality_2013}{EvaluateCore} to evaluate a core set (CS)
selected from an entire collection (EC).
}
\details{
This function is a wrapper around the \code{\link[corehunter]{evaluateCore}}
function of the \code{\link[corehunter]{corehunter}} package.
}
\examples{

####################################
# Use data from R package ccChooser
####################################

library(ccChooser)
data("dactylis_CC")
data("dactylis_EC")

ec <- cbind(genotypes = rownames(dactylis_EC), dactylis_EC[, -1])
ec$genotypes <- as.character(ec$genotypes)
rownames(ec) <- NULL
ec[, c("X1", "X6", "X7")] <- lapply(ec[, c("X1", "X6", "X7")],
                                    function(x) cut(x, breaks = 4))
ec[, c("X1", "X6", "X7")] <- lapply(ec[, c("X1", "X6", "X7")],
                                    function(x) factor(as.numeric(x)))
head(ec)

core <- rownames(dactylis_CC)

quant <- c("X2", "X3", "X4", "X5", "X8")
qual <- c("X1", "X6", "X7")

####################################
# EvaluateCore
####################################

dist.evaluate.core(data = ec, names = "genotypes", quantitative = quant,
                   qualitative = qual, selected = core)


}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[corehunter]{evaluateCore}}
}
