% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.evaluate.core.R
\name{dist.evaluate.core}
\alias{dist.evaluate.core}
\title{Distance Measures}
\usage{
dist.evaluate.core(data, names, quantitative, qualitative, selected)
}
\arguments{
\item{data}{The data as a data frame object. The data frame should possess
one row per individual and columns with the individual names and multiple
trait/character data.}

\item{names}{Name of column with the individual names as a character string}

\item{quantitative}{Name of columns with the quantitative traits as a
character vector.}

\item{qualitative}{Name of columns with the qualitative traits as a
character vector.}

\item{selected}{Character vector with the names of individuals selected in
core collection and present in the \code{names} column.}
}
\value{

}
\description{
Compute Average entry-to-nearest-entry distance, Average
accession-to-nearest-entry distance and Average entry-to-entry distance to
evaluate a core set (CS) selected from an entire collection (EC).
}
\details{
This function is a wrapper around the \code{\link[corehunter]{evaluateCore}}
function of the \code{\link[corehunter]{corehunter}} package.
}
\examples{

}
\seealso{
\code{\link[corehunter]{evaluateCore}}
}
