% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.evaluate.core.R
\name{diversity.evaluate.core}
\alias{diversity.evaluate.core}
\title{Diversity Indices}
\usage{
diversity.evaluate.core(data, names, qualitative, selected, base = 2, R = 1000)
}
\arguments{
\item{data}{The data as a data frame object. The data frame should possess
one row per individual and columns with the individual names and multiple
trait/character data.}

\item{names}{Name of column with the individual names as a character string}

\item{qualitative}{Name of columns with the qualitative traits as a character
vector.}

\item{selected}{Character vector with the names of individuals selected in
core collection and present in the \code{names} column.}

\item{base}{The logarithm base to be used for computation of Shannon-Weaver
Diversity Index (\mjseqn{I}). Default is 2.}

\item{R}{The number of bootstrap replicates.}
}
\value{
A data frame with the following columns. \item{Trait}{The qualitative
  trait.} \item{EC_No.Classes}{The number of classes in the trait for EC.}
  \item{EC_D}{The Simpson's Index (\mjseqn{D}) for EC.} \item{EC_H}{Nei's
  Diversity Index (\mjseqn{H}) for EC.} \item{EC_H.max}{Maximum Nei's
  Diversity Index (\mjseqn{H_{max}}) for EC.} \item{EC_D.inv}{Simpson's
  Reciprocal Index (\mjseqn{D_{R}}) for EC.} \item{EC_H.rel}{Relative Nei's
  Diversity Index (\mjseqn{H'}) for EC.} \item{EC_I}{Shannon-Weaver Diversity
  Index (\mjseqn{I}) for EC.} \item{EC_I.max}{Maximum Shannon-Weaver
  Diversity Index (\mjseqn{I_{max}}) for EC.} \item{EC_I.rel}{Relative
  Shannon-Weaver Diversity Index (\mjseqn{I'}) for EC.}
  \item{EC_D.Mc}{McIntosh Diversity Index (\mjseqn{D_{Mc}}) for EC.}
  \item{CS_No.Classes}{The number of classes in the trait for CS.}
  \item{CS_D}{The Simpson's Index (\mjseqn{D}) for CS.} \item{CS_H}{Nei's
  Diversity Index (\mjseqn{H}) for CS.} \item{CS_H.max}{Maximum Nei's
  Diversity Index (\mjseqn{H_{max}}) for CS.} \item{CS_D.inv}{Simpson's
  Reciprocal Index (\mjseqn{D_{R}}) for CS.} \item{CS_H.rel}{Relative Nei's
  Diversity Index (\mjseqn{H'}) for CS.} \item{CS_I}{Shannon-Weaver Diversity
  Index (\mjseqn{I}) for CS.} \item{CS_I.max}{Maximum Shannon-Weaver
  Diversity Index (\mjseqn{I_{max}}) for CS.} \item{CS_I.rel}{Relative
  Shannon-Weaver Diversity Index (\mjseqn{I'}) for CS.}
  \item{CS_D.Mc}{McIntosh Diversity Index (\mjseqn{D_{Mc}}) for CS.}.
}
\description{
Compute the following diversity indices and perform corresponding statistical
tests to compare the phenotypic diversity for qualitative traits between
entire collection (EC) and core set (CS). \itemize{ \item{Simpson's and
related indices} \itemize{ \item{Simpson's Index (\mjseqn{d})
\insertCite{simpson_measurement_1949,peet_measurement_1974}{EvaluateCore}}
\item{Simpson's Index of Diversity or Gini's Diversity Index or Gini-Simpson
Index or Nei's Diversity Index or Nei's Variation Index (\mjseqn{D})
\insertCite{gini_variabilita_1912,gini_variabilita_1912-2,greenberg_measurement_1956,berger_diversity_1970,nei_analysis_1973,peet_measurement_1974}{EvaluateCore}}
\item{Maximum Simpson's Index of Diversity or Maximum Nei's
Diversity/Variation Index (\mjseqn{D_{max}})
\insertCite{hennink_interpretation_1990}{EvaluateCore}} \item{Simpson's
Reciprocal Index or Hill's \mjseqn{N_{2}} (\mjseqn{D_{R}})
\insertCite{williams_patterns_1964,hill_diversity_1973}{EvaluateCore}}
\item{Relative Simpson's Index of Diversity or Relative Nei's
Diversity/Variation Index (\mjseqn{D'})
\insertCite{hennink_interpretation_1990}{EvaluateCore}} }
\item{Shannon-Weaver and related indices} \itemize{ \item{Shannon or
Shannon-Weaver or Shannon-Weiner Diversity Index (\mjseqn{H})
\insertCite{shannon_mathematical_1949,peet_measurement_1974}{EvaluateCore}}
\item{Maximum Shannon-Weaver Diversity Index (\mjseqn{H_{max}})
\insertCite{hennink_interpretation_1990}{EvaluateCore}} \item{Relative
Shannon-Weaver Diversity Index or Shannon Equitability Index (\mjseqn{H'})
\insertCite{hennink_interpretation_1990}{EvaluateCore}} } \item{McIntosh
Diversity Index} \itemize{ \item{McIntosh Diversity Index (\mjseqn{D_{Mc}})
\insertCite{mcintosh_index_1967,peet_measurement_1974}{EvaluateCore}} } }
\loadmathjax
}
\section{Details}{
 The diversity indices and the corresponding statistical
  tests implemented in \code{diversity.evaluate.core} are as follows.

  \subsection{Simpson's and related indices}{Simpson's index (\mjseqn{d})
  which estimates the probability that two accessions randomly selected will
  belong to the same phenotypic class of a trait, is computed as follows
  \insertCite{simpson_measurement_1949,peet_measurement_1974}{EvaluateCore}.

  \mjsdeqn{d = \sum_{i = 1}^{k}p_{i}^{2}}

  Where, \mjseqn{p_{i}} denotes the proportion/fraction/frequency of
  accessions in the \mjseqn{i}th phenotypic class for a trait and \mjseqn{k}
  is the number of phenotypic classes for the trait.

  The value of \mjseqn{d} can range from 0 to 1 with 0 representing maximum
  diversity and 1, no diversity.

  \mjseqn{d} is subtracted from 1 to give Simpson's index of diversity
  (\mjseqn{D})
  \insertCite{greenberg_measurement_1956,berger_diversity_1970,peet_measurement_1974,hennink_interpretation_1990}{EvaluateCore}
   originally suggested by
  \insertCite{gini_variabilita_1912,gini_variabilita_1912-2;textual}{EvaluateCore}
   and described in literature as Gini's diversity index or Gini-Simpson
  index. It is the same as Nei's diversity index or Nei's variation index
  \insertCite{nei_analysis_1973,hennink_interpretation_1990}{EvaluateCore}.
  Greater the value of \mjseqn{D}, greater the diversity with a range from 0
  to 1.

  \mjsdeqn{D = 1 - d}

  The maximum value of \mjseqn{D}, \mjseqn{D_{max}} occurs when accessions
  are uniformly distributed across the phenotypic classes and is computed as
  follows \insertCite{hennink_interpretation_1990}{EvaluateCore}.

  \mjsdeqn{D_{max} = 1 - \frac{1}{k}}

  Reciprocal of \mjseqn{d} gives the Simpson's reciprocal index
  (\mjseqn{D_{R}})
  \insertCite{williams_patterns_1964,hennink_interpretation_1990}{EvaluateCore}
   and can range from 1 to \mjseqn{k}. This was also described in
  \insertCite{hill_diversity_1973;textual}{EvaluateCore} as (\mjseqn{N_{2}}).

  \mjsdeqn{D_{R} = \frac{1}{d}}

  Relative Simpson's index of diversity or Relative Nei's diversity/variation
  index (\mjseqn{H'}) \insertCite{hennink_interpretation_1990}{EvaluateCore}
  is defined as follows \insertCite{peet_measurement_1974}{EvaluateCore}.

  \mjsdeqn{D' = \frac{D}{D_{max}}}

  Differences in Simpson's diversity index for qualitative traits of EC and
  CS can be tested by Hutcheson t-test
  \insertCite{hutcheson_test_1970}{EvaluateCore}.

  The t statistic is computed as follows.

  \mjsdeqn{t = \frac{H_{EC} - H_{CS}}{\sqrt{V_{H_{EC}} + V_{H_{CS}}}}}

  Where, the variance of \mjseqn{H} (\mjseqn{V_{H}}) is,

  \mjsdeqn{V_{H} = \frac{\sum_{i=1}^{k}n_{i}(\log{n_{i}})^{2} -
  \frac{(\sum_{i=1}^{k}\log{n_{i}})^2}{N}}{N^{2}}}

  }

  \subsection{Shannon-Weaver and related indices}{An index of information
  \mjseqn{H}, was described by
  \insertCite{shannon_mathematical_1949;textual}{EvaluateCore} as follows.

  \mjsdeqn{H = -\sum_{i=1}^{k}p_{i} \log_{2}(p_{i})}

  \mjseqn{H} is described as Shannon or Shannon-Weaver or Shannon-Weiner
  diversity index in literature.

  Alternatively, \mjseqn{H} is also computed using natural logarithm instead
  of logarithm to base 2.

  \mjsdeqn{H = -\sum_{i=1}^{k}p_{i} \ln(p_{i})}

  The maximum value of \mjseqn{H} (\mjseqn{H_{max}}) is \mjseqn{\ln(k)}. This
  value occurs when each phenotypic class for a trait has the same proportion
  of accessions.

  \mjsdeqn{H_{max} = \log_{2}(k)} OR \mjsdeqn{H_{max} = \ln(k)}

  The relative Shannon-Weaver diversity index or Shannon equitability index
  (\mjseqn{H'}) is the Shannon diversity index (\mjseqn{I}) divided by the
  maximum diversity (\mjseqn{H_{max}}).

  \mjsdeqn{H' = \frac{H}{H_{max}}}}

  \subsection{McIntosh Diversity Index}{A similar index of diversity was
  described by \insertCite{mcintosh_index_1967;textual}{EvaluateCore} as
  follows (\mjseqn{D_{Mc}}) \insertCite{peet_measurement_1974}{EvaluateCore}.

  \mjsdeqn{D_{Mc} = \frac{N - \sqrt{\sum_{i=1}^{k}n_{i}^2}}{N - \sqrt{N}}}

  Where, \mjseqn{n_{i}} denotes the number of accessions in the \mjseqn{i}th
  phenotypic class for a trait and \mjseqn{N} is the total number of
  accessions so that \mjseqn{p_{i} = {n_{i}}/{N}}.}
}

\examples{

data("cassava_CC")
data("cassava_EC")

ec <- cbind(genotypes = rownames(cassava_EC), cassava_EC)
ec$genotypes <- as.character(ec$genotypes)
rownames(ec) <- NULL

core <- rownames(cassava_CC)

quant <- c("NMSR", "TTRN", "TFWSR", "TTRW", "TFWSS", "TTSW", "TTPW", "AVPW",
           "ARSR", "SRDM")
qual <- c("CUAL", "LNGS", "PTLC", "DSTA", "LFRT", "LBTEF", "CBTR", "NMLB",
          "ANGB", "CUAL9M", "LVC9M", "TNPR9M", "PL9M", "STRP", "STRC",
          "PSTR")

ec[, qual] <- lapply(ec[, qual],
                     function(x) factor(as.factor(x)))

diversity.evaluate.core(data = ec, names = "genotypes",
                        qualitative = qual, selected = core)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[psych:misc]{shannon}}, \code{\link[vegan]{diversity}},
  \code{\link[boot]{boot}}
}
