% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.evaluate.core.R
\name{pca.evaluate.core}
\alias{pca.evaluate.core}
\title{Principal Component Analysis}
\usage{
pca.evaluate.core(data, names, quantitative, selected, center = TRUE,
  scale = TRUE, npc.plot = 6)
}
\arguments{
\item{data}{The data as a data frame object. The data frame should possess
one row per individual and columns with the individual names and multiple
trait/character data.}

\item{names}{Name of column with the individual names as a character string}

\item{quantitative}{Name of columns with the quantitative traits as a
character vector.}

\item{selected}{Character vector with the names of individuals selected in
core collection and present in the \code{names} column.}
}
\value{

}
\description{
Compute Principal Component Analysis Statistics
\insertCite{mardia_multivariate_1979}{EvaluateCore} to compare the
probability distributions of quantitative traits between entire collection
(EC) and core set (CS).
}
\examples{

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[stats]{prcomp}}
}
