% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentdiff.evaluate.core.R
\name{percentdiff.evaluate.core}
\alias{percentdiff.evaluate.core}
\title{Percentage Difference of Means and Variances}
\usage{
percentdiff.evaluate.core(data, names, quantitative, selected,
  alpha = 0.05)
}
\arguments{
\item{data}{The data as a data frame object. The data frame should possess
one row per individual and columns with the individual names and multiple
trait/character data.}

\item{names}{Name of column with the individual names as a character string}

\item{quantitative}{Name of columns with the quantitative traits as a
character vector.}

\item{selected}{Character vector with the names of individuals selected in
core collection and present in the \code{names} column.}

\item{alpha}{Type I error probability (Significance level) of difference.}
}
\value{

}
\description{
Compute the following differences between the entire collection (EC) and core
set (CS). \itemize{ \item Percentage of significant differences of mean
(\ifelse{html}{\out{<em>MD\%<sub>Hu</sub></em>}}{\eqn{MD\%_{Hu}}})
\insertCite{hu_methods_2000}{EvaluateCore} \item Percentage of significant
differences of variance
(\ifelse{html}{\out{<em>VD\%<sub>Hu</sub></em>}}{\eqn{VD\%_{Hu}}})
\insertCite{hu_methods_2000}{EvaluateCore} \item Average of absolute
differences between means
(\ifelse{html}{\out{<em>MD\%<sub>Kim</sub></em>}}{\eqn{MD\%_{Kim}}})
\insertCite{kim_powercore:_2007}{EvaluateCore} \item Average of absolute
differences between variances
(\ifelse{html}{\out{<em>VD\%<sub>Kim</sub></em>}}{\eqn{VD\%_{Kim}}})
\insertCite{kim_powercore:_2007}{EvaluateCore} }
}
\details{
The differences are computed as follows.

\ifelse{html}{\out{<p style="text-align: center;"><em>MD\%<sub>Hu</sub></em>
= <big>(</big><sup>S<sub>t</sub></sup> &frasl; <sub>n</sub><big>)</big>
&times; 100</p>}}{\deqn{MD\%_{Hu} = \left ( \frac{S_{t}}{n} \right ) \times
100}}

Where \ifelse{html}{\out{<em>S<sub>t</sub></em>}}{\eqn{S_{t}}} is the number
of traits with a significant difference between the means of the EC and the
CS and \ifelse{html}{\out{<em>n</em>}}{\eqn{n}} is the total number of
traits.

\ifelse{html}{\out{<p style="text-align: center;"><em>VD\%<sub>Hu</sub></em>
= <big>(</big><sup>S<sub>t</sub></sup> &frasl; <sub>n</sub><big>)</big>
&times; 100</p>}}{\deqn{VD\%_{Hu} = \left ( \frac{S_{t}}{n} \right ) \times
100}}

Where \ifelse{html}{\out{<em>S<sub>t</sub></em>}}{\eqn{S_{t}}} is the number
of traits with a significant difference between the variances of the EC and
the CS and \ifelse{html}{\out{<em>n</em>}}{\eqn{n}} is the total number of
traits.

\ifelse{html}{\out{<p style="text-align: center;"><em>MD\%<sub>Kim</sub> =
<sup>1</sup>&frasl;<sub>n</sub> &sum;<sup>n</sup><sub style="line-height:
1.8; margin-left: -1ex;">i=1</sub> <big>[</big>
<sup>|M<sub>EC<sub>i</sub></sub> &minus; M<sub>CS<sub>i</sub></sub>|</sup>
&frasl; <sub>M<sub>CS<sub>i</sub></sub></sub> <big>]</big>  &times;
100</em></p>}}{\deqn{MD\%_{Kim} = \frac{1}{n}\sum_{i=1}^{n} \frac{\left |
M_{EC_{i}}-M_{CS_{i}} \right |}{M_{CS_{i}}}}}

Where
\ifelse{html}{\out{<em>M<sub>EC<sub>i</sub></sub></em>}}{\eqn{M_{EC_{i}}}} is
the mean of the EC for the \ifelse{html}{\out{<em>i</em>}}{\eqn{i}}th trait,
\ifelse{html}{\out{<em>M<sub>CS<sub>i</sub></sub></em>}}{\eqn{M_{CS_{i}}}} is
the mean of the CS for the \ifelse{html}{\out{<em>i</em>}}{\eqn{i}}th trait
and \ifelse{html}{\out{<em>n</em>}}{\eqn{n}} is the total number of traits.

\ifelse{html}{\out{<p style="text-align: center;"><em>VD\%<sub>Kim</sub> =
<sup>1</sup>&frasl;<sub>n</sub> &sum;<sup>n</sup><sub style="line-height:
1.8; margin-left: -1ex;">i=1</sub> <big>[</big>
<sup>|V<sub>EC<sub>i</sub></sub> &minus; V<sub>CS<sub>i</sub></sub>|</sup>
&frasl; <sub>V<sub>CS<sub>i</sub></sub></sub> <big>]</big>  &times;
100</em></p>}}{\deqn{VD\%_{Kim} = \frac{1}{n}\sum_{i=1}^{n} \frac{\left |
V_{EC_{i}}-V_{CS_{i}} \right |}{V_{CS_{i}}}}}

Where
\ifelse{html}{\out{<em>V<sub>EC<sub>i</sub></sub></em>}}{\eqn{V_{EC_{i}}}} is
the variance of the EC for the \ifelse{html}{\out{<em>i</em>}}{\eqn{i}}th
trait,
\ifelse{html}{\out{<em>V<sub>CS<sub>i</sub></sub></em>}}{\eqn{V_{CS_{i}}}} is
the variance of the CS for the \ifelse{html}{\out{<em>i</em>}}{\eqn{i}}th
trait and \ifelse{html}{\out{<em>n</em>}}{\eqn{n}} is the total number of
traits.
}
\examples{

####################################
# Use data from R package ccChooser
####################################

library(ccChooser)
data("dactylis_CC")
data("dactylis_EC")

ec <- cbind(genotypes = rownames(dactylis_EC), dactylis_EC[, -1])
ec$genotypes <- as.character(ec$genotypes)
rownames(ec) <- NULL
ec[, c("X1", "X6", "X7")] <- lapply(ec[, c("X1", "X6", "X7")],
                                    function(x) cut(x, breaks = 4))
ec[, c("X1", "X6", "X7")] <- lapply(ec[, c("X1", "X6", "X7")],
                                    function(x) factor(as.numeric(x)))
head(ec)

core <- rownames(dactylis_CC)

quant <- c("X2", "X3", "X4", "X5", "X8")
qual <- c("X1", "X6", "X7")

####################################
# EvaluateCore
####################################

percentdiff.evaluate.core(data = ec, names = "genotypes",
                          quantitative = quant, selected = core)


}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[EvaluateCore]{snk.evaluate.core}},
  \code{\link[EvaluateCore]{snk.evaluate.core}}
}
