% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signtest.evaluate.core.R
\name{signtest.evaluate.core}
\alias{signtest.evaluate.core}
\title{Sign Test}
\usage{
signtest.evaluate.core(data, names, quantitative, selected)
}
\arguments{
\item{data}{The data as a data frame object. The data frame should possess
one row per individual and columns with the individual names and multiple
trait/character data.}

\item{names}{Name of column with the individual names as a character string}

\item{quantitative}{Name of columns with the quantitative traits as a
character vector.}

\item{selected}{Character vector with the names of individuals selected in
core collection and present in the \code{names} column.}
}
\value{

}
\description{
Test difference between means and variances of entire collection (EC) and
core set (CS) for quantitative traits by Sign test
\insertCite{basigalup_development_1995,tai_core_2001}{EvaluateCore}.
}
\examples{

}
\references{
\insertAllCited{}
}
