% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.evaluate.core.R
\name{ttest.evaluate.core}
\alias{ttest.evaluate.core}
\title{Student's t Test}
\usage{
ttest.evaluate.core(data, names, quantitative, selected)
}
\arguments{
\item{data}{The data as a data frame object. The data frame should possess
one row per individual and columns with the individual names and multiple
trait/character data.}

\item{names}{Name of column with the individual names as a character string}

\item{quantitative}{Name of columns with the quantitative traits as a
character vector.}

\item{selected}{Character vector with the names of individuals selected in
core collection and present in the \code{names} column.}
}
\value{

}
\description{
Test difference between means of entire collection (EC) and core set (CS) for
quantitative traits by Student's t test
\insertCite{student_probable_1908}{EvaluateCore}.
}
\examples{

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[stats]{t.test}}
}
